[c++] Upgrade rapidjson to v1.1
v1.1 addresses warning from Clang. Also, fix some issues in VC++ that the
upgrade exposed, specifically:

* Unreachable code warnings
* Conflicts due to mix/max macros in windef.h
* Warning due to converting enum to uint16_t

+++ cpp/inc/bond/comm/wire_protocol.h
@@ -98,7 +98,7 @@ public:
     {
         auto stream = StreamFactory::Create();
 
-        ::bond::Apply< ::bond::Marshaler>(payload, stream, requestProtocol);
+        ::bond::Apply< ::bond::Marshaler>(payload, stream, static_cast<uint16_t>(requestProtocol));
 
         std::vector<blob> buffers;
         stream.GetBuffers(buffers);
